+++ infra/docker-compose.yaml
@@ -31,5 +31,11 @@ services:
     networks:
       - healenium
 
+  selector-imitator:
+    image: healenium/hlm-selector-imitator:1
+    restart: on-failure
+    ports:
+      - "8000:8000"
+
 networks:
   healenium:
