+++ src/main/java/com/epam/healenium/healenium_proxy/config/HealeniumProxyAppConfig.java
@@ -1,0 +1,15 @@
+package com.epam.healenium.healenium_proxy.config;
+
+import com.epam.healenium.healenium_proxy.handler.HealeniumProxyHttpHandler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.HttpRequestHandler;
+
+@Configuration
+public class HealeniumProxyAppConfig {
+
+    @Bean(name = "/**")
+    public HttpRequestHandler httpRequestHandler() {
+        return new HealeniumProxyHttpHandler();
+    }
+}
