+++ src/test/java/com/epam/healenium/healenium_proxy/HealeniumProxyApplicationTests.java
@@ -1,4 +1,4 @@
-package com.epam.healenium.HealeniumProxy;
+package com.epam.healenium.healenium_proxy;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
